further reading link
This is an excellent article. I notice on legacy challenges links for further reading are included to various 3rd party sites. I hope this meets the style guides. This article really crystalized things for me when I was first started seriously studying the concept. While it was written in 2010 it seems relevant and gives some context.